refactor: test_cancellation subprocess start
